GF-5[refactor]: 쿠폰생성 DTO 분리
