+++ src/main/java/com/example/screen_golf/coupon/domain/UserCoupon.java
@@ -20,7 +20,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -95,36 +94,4 @@ public class UserCoupon {
 			now.isBefore(validTo);
 	}
 
-	/**
-	 *=====================================================
-	 * 					UserCouponDto
-	 * =====================================================
-	 */
-
-	@Getter
-	@NoArgsConstructor
-	public static class UserCouponCreateRequest {
-		private Long userId;         // 사용자 ID
-		private String couponCode;   // 쿠폰 코드
-		private String name;         // 쿠폰 이름
-		private Integer discountAmount; // 할인 금액
-		private LocalDateTime validFrom; // 쿠폰 유효 시작일
-		private LocalDateTime validTo;   // 쿠폰 유효 종료일
-	}
-
-	@Builder
-	@Getter
-	@AllArgsConstructor
-	@NoArgsConstructor
-	public static class UserCouponCreateResponse {
-
-		private Long userCouponId;   // 생성된 쿠폰 ID
-		private Long userId;         // 사용자 ID
-		private String couponCode;   // 쿠폰 코드
-		private String getCouponName;         // 쿠폰 이름
-		private Integer discountAmount; // 할인 금액
-		private LocalDateTime validFrom; // 쿠폰 유효 시작일
-		private LocalDateTime validTo;   // 쿠폰 유효 종료일
-		private LocalDateTime createdAt; // 생성일시
-	}
 }
